Added bigint support.

+++ resources/db/test-db/migrations/deploy/create-twitter-tweets-table.sql
@@ -3,8 +3,8 @@
 BEGIN;
 
 CREATE TABLE twitter.tweets (
-  id integer PRIMARY KEY,
-  user_id integer REFERENCES twitter.users(id) ON DELETE CASCADE,
+  id bigint PRIMARY KEY,
+  user_id bigint REFERENCES twitter.users(id) ON DELETE CASCADE,
   retweeted boolean NOT NULL DEFAULT false,
   text text NOT NULL,
   created_at timestamp with time zone NOT NULL DEFAULT now(),
